// apps/soundbyte-web/src/types/soundcloud-playlist.ts
export type SoundCloudProduct = {
  id: string;
  name: string;
};

export type SoundCloudSubscription = {
  product: SoundCloudProduct;
};

export type SoundCloudUser = {
  avatar_url: string;
  id: number;
  urn: string;
  kind: "user";
  permalink_url: string;
  uri: string;
  username: string;
  permalink: string;
  created_at: string;
  last_modified: string;
  first_name: string;
  last_name: string;
  full_name: string;
  city: string | null;
  description: string | null;
  country: string | null;
  track_count: number;
  public_favorites_count: number;
  reposts_count: number;
  followers_count: number;
  followings_count: number;
  plan: string;
  myspace_name: string | null;
  discogs_name: string | null;
  website_title: string | null;
  website: string | null;
  comments_count: number;
  online: boolean;
  likes_count: number;
  playlist_count: number;
  subscriptions: SoundCloudSubscription[];
};

export type SoundCloudTrack = {
  kind: "track";
  id: number;
  urn: string;
  created_at: string;
  duration: number;
  commentable: boolean;
  comment_count: number;
  sharing: string;
  tag_list: string;
  streamable: boolean;
  embeddable_by: string;
  purchase_url: string | null;
  purchase_title: string | null;
  genre: string | null;
  title: string;
  description: string | null;
  label_name: string | null;
  release: string | null;
  key_signature: string | null;
  isrc: string | null;
  bpm: number | null;
  release_year: number | null;
  release_month: number | null;
  release_day: number | null;
  license: string;
  uri: string;
  user: SoundCloudUser;
  permalink_url: string;
  artwork_url: string | null;
  stream_url: string | null;
  download_url: string | null;
  waveform_url: string | null;
  available_country_codes: string[] | null;
  secret_uri: string | null;
  user_favorite?: boolean;
  user_playback_count?: number;
  playback_count?: number;
  download_count?: number;
  favoritings_count?: number;
  reposts_count?: number;
  downloadable: boolean | null;
  access?: string;
  policy?: string | null;
  monetization_model?: string | null;
  metadata_artist?: string | null;
};

export type SoundCloudPlaylist = {
  duration: number;
  genre: string;
  release_day: number | null;
  permalink: string;
  permalink_url: string;
  release_month: number | null;
  release_year: number | null;
  description: string | null;
  uri: string;
  label_name: string | null;
  label_id: number | null;
  label: string | null;
  tag_list: string;
  track_count: number;
  user_id: number;
  user_urn: string;
  last_modified: string;
  license: string;
  user: SoundCloudUser;
  playlist_type: string;
  type: string;
  id: number;
  urn: string;
  downloadable: boolean | null;
  likes_count: number;
  repost_count: number;
  sharing: string;
  created_at: string;
  release: string | null;
  tags: string;
  kind: "playlist";
  title: string;
  purchase_title: string | null;
  ean: string | null;
  streamable: boolean;
  embeddable_by: string;
  artwork_url: string | null;
  purchase_url: string | null;
  tracks_uri: string;
  tracks: SoundCloudTrack[];
};

export type SoundCloudPlaylistResponse = SoundCloudPlaylist[];